first refactor: location db queries
